pokerServiceImpl setup + changed size_t to uin32t for .proto integration

+++ server/include/Entities/Action.h
@@ -10,13 +10,13 @@ using ActionType = Enums::ActionType;
 class Action {
 private:
     string playerId;
-    size_t betAmount;
+    uint32_t betAmount;
 
 public:
-    Action(string id, size_t amount = 0) : playerId(id), betAmount(amount) {}
+    Action(string id, uint32_t amount = 0) : playerId(id), betAmount(amount) {}
 
     const string& getPlayerId() { return playerId; }
-    const size_t& getAmount() { return betAmount; }
+    const uint32_t& getAmount() { return betAmount; }
     
     virtual ~Action() {}
     virtual ActionType getActionType() const = 0;
