+++ server/include/Entities/Actions/CallAction.h
@@ -5,7 +5,7 @@
 
 class CallAction : public Action {
 public:
-    CallAction(string id, size_t amount) : Action(id, amount) {}
+    CallAction(string id, uint32_t amount) : Action(id, amount) {}
     ActionType getActionType() const override { return ActionType::CALL; }
 };
