+++ server/include/Entities/Actions/PostSmallAction.h
@@ -5,7 +5,7 @@
 
 class PostSmallAction : public Action {
 public:
-    PostSmallAction(string id, size_t amount) : Action(id, amount) {}
+    PostSmallAction(string id, uint32_t amount) : Action(id, amount) {}
     ActionType getActionType() const override { return ActionType::POST_SMALL; }
 };
